services/check_code/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
google-genai
orjson
python-jose==3.4.0
ollama
faiss-cpu
scikit-learn
transformers
torch
sentencepiece
